move af device widget
